cleaned up the folders (deleted the test.py files), added in a backend that functions correctly with both the frontend and db. Changed frontend to have functionality with what i had made in html.
Commented out everything in main.py in the backend folder, as all of the code i wrote now uses app.py. Left main.py available tho just in case we run into problems.

To run the full app: open 2 split terminals. In one, cd to the directory of app.py (backend). run app.py. The terminal should give you confirmation.
In the other, cd to the directory of App.jsx. Use npm start to run the React app.

As far as I know, you have to run the flask file first, and then the react app.

The DB is 1) a test db and 2) local, so the app will only work if you insert your own db url in the config.py file until we get a more permamanent db running.